feat: update deleteCloudtty

+++ src/otomi-stack.ts
@@ -720,9 +720,8 @@ export default class OtomiStack {
   async deleteCloudtty(data: Cloudtty) {
     const cloudttys = this.db.getCollection('cloudttys') as Array<Cloudtty>
     const cloudtty = cloudttys.find((c) => c.emailNoSymbols === data.emailNoSymbols) as Cloudtty
-    if (cloudtty.id) await k8sdelete(data)
-    if (!(await checkPodExists('team-admin', `tty-${data.emailNoSymbols}`)))
-      return this.db.deleteItem('cloudttys', { id: cloudtty.id })
+    if (await checkPodExists('team-admin', `tty-${data.emailNoSymbols}`)) await k8sdelete(data)
+    if (cloudtty.id) return this.db.deleteItem('cloudttys', { id: cloudtty.id })
   }
 
   async saveCloudttys(): Promise<void> {
